w2-gm/requirements.txt
websockets==11.0.3
websocket-client==1.6.2
docker